scripts/requirements.txt
websockets>=12.0
pyserial>=3.5